LRQA-42795 Create BaseGitRemote
