+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/GitWorkingDirectory.java
@@ -1687,149 +1687,14 @@ public class GitWorkingDirectory {
 		throw new RuntimeException("Unable to run: git status");
 	}
 
-	public static class Remote implements Comparable<Remote> {
+	public static class Remote extends BaseGitRemote {
 
-		@Override
-		public int compareTo(Remote otherGitRemote) {
-			int result = _name.compareTo(otherGitRemote._name);
-
-			if (result != 0) {
-				return result;
-			}
-
-			return _fetchRemoteURL.compareTo(otherGitRemote._fetchRemoteURL);
-		}
-
-		public GitWorkingDirectory getGitWorkingDirectory() {
-			return _gitWorkingDirectory;
-		}
-
-		public String getHostname() {
-			return _hostname;
-		}
-
-		public String getName() {
-			return _name;
-		}
-
-		public String getPushRemoteURL() {
-			if (_pushRemoteURL != null) {
-				return _pushRemoteURL;
-			}
-
-			return _fetchRemoteURL;
-		}
-
-		public String getRemoteURL() {
-			return _fetchRemoteURL;
-		}
-
-		public String getRepositoryName() {
-			return _repositoryName;
-		}
-
-		public String getUsername() {
-			return _username;
-		}
-
-		public String toString() {
-			return JenkinsResultsParserUtil.combine(
-				getName(), " (", getRemoteURL(), ")");
-		}
-
-		private Remote(
+		protected Remote(
 			GitWorkingDirectory gitWorkingDirectory,
 			String[] remoteInputLines) {
 
-			_gitWorkingDirectory = gitWorkingDirectory;
-
-			if (remoteInputLines.length != 2) {
-				throw new IllegalArgumentException(
-					"\"remoteInputLines\" must be an array of 2 strings");
-			}
-
-			if (remoteInputLines[0].equals(remoteInputLines[1])) {
-				throw new IllegalArgumentException(
-					JenkinsResultsParserUtil.combine(
-						"\"remoteInputLines[0]\" and ",
-						"\"remoteInputLines[1]\" are identical: ",
-						remoteInputLines[0]));
-			}
-
-			if ((remoteInputLines[0] == null) ||
-				(remoteInputLines[1] == null)) {
-
-				throw new IllegalArgumentException(
-					"Neither \"remoteInputLines[0]\" nor " +
-						"\"remoteInputLines[1]\" may be NULL: " +
-							Arrays.toString(remoteInputLines));
-			}
-
-			String name = null;
-			String fetchRemoteURL = null;
-			String pushRemoteURL = null;
-
-			for (String remoteInputLine : remoteInputLines) {
-				Matcher matcher = _remotePattern.matcher(remoteInputLine);
-
-				if (!matcher.matches()) {
-					throw new IllegalArgumentException(
-						"Invalid Git remote input line " + remoteInputLine);
-				}
-
-				if (name == null) {
-					name = matcher.group("name");
-				}
-
-				String remoteURL = matcher.group("remoteURL");
-				String type = matcher.group("type");
-
-				if ((fetchRemoteURL == null) && type.equals("fetch")) {
-					fetchRemoteURL = remoteURL;
-				}
-
-				if ((pushRemoteURL == null) && type.equals("push")) {
-					pushRemoteURL = remoteURL;
-				}
-			}
-
-			_fetchRemoteURL = fetchRemoteURL;
-			_name = name;
-			_pushRemoteURL = pushRemoteURL;
-
-			Matcher remoteURLMatcher = _remoteURLMultiPattern.matches(
-				_fetchRemoteURL);
-
-			if (remoteURLMatcher == null) {
-				throw new RuntimeException(
-					JenkinsResultsParserUtil.combine(
-						"fetch remote URL ", _fetchRemoteURL,
-						" is not a valid remote URL"));
-			}
-
-			_hostname = remoteURLMatcher.group("hostname");
-			_username = remoteURLMatcher.group("username");
-			_repositoryName = remoteURLMatcher.group("repositoryName");
+			super(gitWorkingDirectory, remoteInputLines);
 		}
-
-		private static final Pattern _remotePattern = Pattern.compile(
-			JenkinsResultsParserUtil.combine(
-				"(?<name>[^\\s]+)[\\s]+(?<remoteURL>[^\\s]+)[\\s]+\\(",
-				"(?<type>[^\\s]+)\\)"));
-		private static final MultiPattern _remoteURLMultiPattern =
-			new MultiPattern(
-				"git@(?<hostname>[^:]+):(?<username>[^/]+)" +
-					"/(?<repositoryName>[^\\.^\\s]+)(\\.git)?+\\s*",
-				"https://(?<hostname>[^/]+)/(?<username>[^/]+)" +
-					"/(?<repositoryName>[^\\.^\\s]+)(\\.git)?+\\s*");
-
-		private final String _fetchRemoteURL;
-		private final GitWorkingDirectory _gitWorkingDirectory;
-		private final String _hostname;
-		private final String _name;
-		private final String _pushRemoteURL;
-		private final String _repositoryName;
-		private final String _username;
 
 	}
